requirements.txt
pyqt5-tools==5.15.9.3.3
pyqt5==5.15.10
pillow==10.1.0
mysql-connector-python==8.2.0
auto-py-to-exe==2.42.0